privatekube/privatekube/kfp/components_src/claim/allocate/src/requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
fsspec==0.7.4
gcsfs==0.6.2
google-auth==1.18.0
google-auth-oauthlib==0.4.1
idna==2.10
kubernetes==11.0.0
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
urllib3==1.25.9
websocket-client==0.57.0
